% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit_stats.R
\name{calcNlmeStats}
\alias{calcNlmeStats}
\title{Calculate all statistics from dose reponse fit}
\usage{
calcNlmeStats(nlme_model, nlme_data)
}
\arguments{
\item{nlme_model}{the results of \code{\link{fitModelNlmeData}}.}

\item{nlme_data}{the nlme data frame used for fitting \code{\link{prepNlmeData}}.}
}
\value{
a data frame inluding IC50, AUC and RMSE.
}
\description{
Calculate all statistics from dose reponse fit
}
\examples{
data("gdsc_example")
gdsc_example <- removeFailedDrugs(gdsc_example)
gdsc_example <- removeMissingDrugs(gdsc_example)
gdsc_example <- normalizeData(gdsc_example)
gdsc_example <- setConcsForNlme(gdsc_example)
nlme_data <- prepNlmeData(gdsc_example, "COSMIC_ID")
\dontrun{
nlme_model <- fitModelNlmeData(nlme_data, isLargeData = F)
nlme_stats <- calcNlmeStats(nlme_model, nlme_data)
}

}
\seealso{
\code{\link{fitModelNlmeData}}, \code{\link{getModelCoef}}, 
 \code{\link{getIC50Matrix}}
}
