% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit.R
\name{fitModel}
\alias{fitModel}
\title{fitModel}
\usage{
fitModel(gDat, vStart = c(8.886464, 1.495953), bLargeScale = TRUE,
  bSilent = TRUE)
}
\arguments{
\item{gDat}{grouped data frame input}

\item{vStart}{starting parameter values for xmid and scal}

\item{bLargeScale}{equivalent to isLargeData in fitModelNlmeData}

\item{bSilent}{verbose output or not}
}
\value{
a fitted nlme model
}
\description{
Used internally by fitModelNlmeData
}
