% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit.R
\name{fitModelNlmeData}
\alias{fitModelNlmeData}
\title{Fit an nlme dose reponse model to GDSC data.}
\usage{
fitModelNlmeData(nlme_data, isLargeData = T)
}
\arguments{
\item{nlme_data}{a data frame as specified by \code{\link{prepNlmeData}}}

\item{isLargeData}{logical. Default \code{(isLargeData = T)}. Set to \code{F}
if model fitting fails to converge, e.g., 
\code{step halving factor reduced below minimum in PNLS step}

With this \code{isLargeData = TRUE}, the covariance between the position and
scale parameter on the cell line level are assumed to be correlated. This 
further stabilizes the fit. In small bespoke screens this is set to FALSE as
the model otherwise struggles to converge.}
}
\value{
a fitted nlme model
}
\description{
Fit an nlme dose reponse model to GDSC data.
}
\examples{
data("gdsc_example")
gdsc_example <- removeFailedDrugs(gdsc_example)
gdsc_example <- removeMissingDrugs(gdsc_example)
gdsc_example <- normalizeData(gdsc_example)
gdsc_example <- setConcsForNlme(gdsc_example)
nlme_data <- prepNlmeData(gdsc_example, "COSMIC_ID")
\dontrun{
nlme_model <- fitModelNlmeData(nlme_data, isLargeData = F)
}

}
\seealso{
\code{\link{prepNlmeData}}, \code{\link{getModelCoef}}, 
 \code{\link{calcNlmeStats}}.
}
