% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdsc_example}
\alias{gdsc_example}
\title{Example raw data set from the Genomics of Drug Sensitivity in Cancer (GDSC) 
project.}
\format{A data frame with 57987 rows and 24 columns

For more details on the data see \code{vignette("gdsc_raw_data_format")}}
\usage{
gdsc_example
}
\description{
Example raw data set from the Genomics of Drug Sensitivity in Cancer (GDSC) 
project.
}
\keyword{datasets}
