% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdscIC50.R
\docType{package}
\name{gdscIC50}
\alias{gdscIC50}
\alias{gdscIC50-package}
\title{gdscIC50}
\description{
A package to prepare for and fit a non-linear mixed effects dose reponse model 
to the Genomics of Drug Sensitivity in Cancer (GDSC) raw data.
}
