% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit_stats.R
\name{getIC50Matrix}
\alias{getIC50Matrix}
\title{Extract a matrix of IC50 values from a model statistics data frame.}
\usage{
getIC50Matrix(model_stats, drug_identifier = "drug", re_name = T,
  measure = "IC50")
}
\arguments{
\item{model_stats}{the results of \code{\link{calcNlmeStats}}.}

\item{drug_identifier}{a character vector indicating which column of 
nlme_stats to use as the "drug" in the IC50matrix. Normally this is the
 \code{drug} column as used in the model fit formula but you may want to 
 change to another identifier for user friendliness e.g. DRUG_ID or 
 DRUG_NAME. The function checks whether there is a one to one mapping 
 between the chosen identifier and the \code{drug} column before proceeding.
 
 The cell line column is named according to the \code{cl_spec} contained in 
 the \code{model_stats}.}

\item{re_name}{logical. Rename the drug columns to GDSCtools standard 
(Drug_..._IC50). Default is TRUE.}

\item{measure}{character. Default to "IC50" but can also be set to "auc".}
}
\value{
a matrix of IC50s.
}
\description{
Extract a matrix of IC50 values from a model statistics data frame.
}
\examples{
data("gdsc_example")
gdsc_example <- removeFailedDrugs(gdsc_example)
gdsc_example <- removeMissingDrugs(gdsc_example)
gdsc_example <- normalizeData(gdsc_example)
gdsc_example <- setConcsForNlme(gdsc_example)
nlme_data <- prepNlmeData(gdsc_example, "COSMIC_ID")
\dontrun{
nlme_model <- fitModelNlmeData(nlme_data, isLargeData = F)
nlme_stats <- calcNlmeStats(nlme_model, nlme_data)
ic50s <- getIC50Matrix(nlme_stats)
aucs <- getIC50Matrix(nlme_stats, measure = "auc")
}

}
\seealso{
\code{\link{getModelCoef}}
}
