% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit.R
\name{getIntegral}
\alias{getIntegral}
\title{getIntegral}
\usage{
getIntegral(x, xmid, scal)
}
\arguments{
\item{x}{concentration of drug treatment usually on transformed scale. 
See setXFromConc()}

\item{xmid}{location parameter of logistic curve}

\item{scal}{scale parameter of logistic curve}
}
\value{
integral value
}
\description{
getIntegral
}
