% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit_stats.R
\name{getModelCoef}
\alias{getModelCoef}
\title{Get the statistics from a fitted nlme dose reponse model.}
\usage{
getModelCoef(nlme_model, nlme_data)
}
\arguments{
\item{nlme_model}{the results of \code{\link{fitModelNlmeData}}.}

\item{nlme_data}{the input data frame to the model fit.}
}
\value{
a data frame of fitted nlme model statistics - xmid (IC50)
 and scal (slope).
}
\description{
Get the statistics from a fitted nlme dose reponse model.
}
\examples{
data("gdsc_example")
gdsc_example <- removeFailedDrugs(gdsc_example)
gdsc_example <- removeMissingDrugs(gdsc_example)
gdsc_example <- normalizeData(gdsc_example)
gdsc_example <- setConcsForNlme(gdsc_example)
nlme_data <- prepNlmeData(gdsc_example, "COSMIC_ID")
\dontrun{
nlme_model <- fitModelNlmeData(nlme_data, isLargeData = F)
nlme_stats <- calcNlmeStats(nlme_model, nlme_data)
}

}
\seealso{
\code{\link{fitModelNlmeData}}, \code{\link{calcNlmeStats}}.
}
