% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit_prep.R
\name{getTimeStamp}
\alias{getTimeStamp}
\title{Utility function to get a time and data stamp for use in file names.}
\usage{
getTimeStamp()
}
\value{
character string in date format "%d%b%y_%H%M"
}
\description{
Utility function to get a time and data stamp for use in file names.
}
\examples{
time_stamp = getTimeStamp()

}
