% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit.R
\name{logist3}
\alias{logist3}
\title{logist3}
\usage{
logist3(x, xmid, scal)
}
\arguments{
\item{x}{concentration of drug treatment usually on transformed scale. 
See setXFromConc().}

\item{xmid}{location parameter of logistic curve}

\item{scal}{scale parameter of logistic curve}
}
\value{
a function object of class selfStart
}
\description{
The 2 parameter logistic function used to model GDSC dose response data. This
model is described further in published in Vis, D.J. et al. Pharmacogenomics
 2016, 17(7):691-700)
}
