% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit_prep.R
\name{normalizeComboData}
\alias{normalizeComboData}
\title{normalizeComboData}
\usage{
normalizeComboData(myDat, trim = T, neg_control = "NC-1",
  pos_control = "B")
}
\arguments{
\item{myDat}{a GDSC raw data data frame.}

\item{trim}{logical indicating whether to trim normalized values to the range
0 to 1. default \code{(trim = T)}}

\item{neg_control}{The tag used to recognise a negative control well - the 
upper end of the dynamic range.}

\item{pos_control}{The tag used to recognise a positive control well - the 
lower end of the dynamic range.}
}
\description{
Now deprecated - using \code{normalizeData} will return identical results.
}
\details{
\code{normalizeComboData} returns normalized intensities for the drug treated
  wells with respect to controls - column \code{normalized_intensity}.
  
Like \code{normalizeData}, the \code{TAG} column is replaced with 
  \code{lib_drug} and \code{dose} columns, but also \code{treatment} (S or C)
  , \code{DRUG_ID_anch}, \code{CONC_anch} and \code{anchor} columns. 
  Combination treatments (C) are treated as if the library alone is to be 
  fitted.
}
\examples{
data("gdsc_example") # Need a combo example here
gdsc_example <- removeFailedDrugs(gdsc_example)
gdsc_example <- removeMissingDrugs(gdsc_example)
gdsc_example <- normalizeComboData(gdsc_example)
gdsc_example <- setConcsForNlme(gdsc_example)
nlme_data <- prepNlmeData(gdsc_example, "COSMIC_ID")

}
\seealso{
\code{\link{removeFailedDrugs}},  \code{\link{removeMissingDrugs}},
  \code{\link{setConcsForNlme}},  \code{\link{prepNlmeData}}, 
  \code{\link{normalizeData}}
}
