% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit_prep.R
\name{normalizeMultiComboData}
\alias{normalizeMultiComboData}
\title{normalizeMultiComboData}
\usage{
normalizeMultiComboData(
  screen_data,
  trim = T,
  neg_control = "NC-1",
  pos_control = "B"
)
}
\arguments{
\item{screen_data}{a GDSC raw data data frame.}

\item{trim}{logical indicating whether to trim normalized values to the range
0 to 1. default \code{(trim = T)}}

\item{neg_control}{The tag used to recognise a negative control well - the 
upper end of the dynamic range.}

\item{pos_control}{The tag used to recognise a positive control well - the 
lower end of the dynamic range.}
}
\description{
Normalize combination data and condense data for >2 treatments
}
\details{
Some combination treatments have more than 2 drugs. This function will 
condense the treatments into an anchor + a library, e.g., A10, A12, L3, L4
becomes A10_A12 with L3_L4.

The concentration for the analysis is at the moment chosen with the lowest 
anchor/library number taking precendence, so in the above example the
concentration of A10 is used as the concentration of the combined A10_A12 anchor.

This code will run slower than normalizeComboData if used for simple 2 drug 
combinations.
}
\seealso{
\code{\link{normalizeComboData}}
}
