% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit_stats.R
\name{plotResponse}
\alias{plotResponse}
\title{plotResponse}
\usage{
plotResponse(model_stats, cell_line, drug_identifier)
}
\arguments{
\item{model_stats}{dataframe of fitted values as produced by \code{calcNlmeStats}}

\item{cell_line}{as identified from the \code{CL} column of the model stats data frame.}

\item{drug_identifier}{character string as identified by the \code{drug} column of the 
model stats data frame.}
}
\value{
ggplot of dose response curve.
}
\description{
Plot dose reponse curve.
}
\examples{
\dontrun{
nlme_model <- fitModelNlmeData(nlme_data, isLargeData = F)
nlme_stats <- calcNlmeStats(nlme_model, nlme_data)
plotResponse(model_stats = nlme_stats, 
             cell_line = 1503364,
             drug_identifier = "1032")
}
}
