% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit_prep.R
\name{prepNlmeData}
\alias{prepNlmeData}
\title{Converts GDSC normalized-data data frame to input format for nlme curve fit.}
\usage{
prepNlmeData(normalized_data, include_combos = F)
}
\arguments{
\item{normalized_data}{a GDSC normalized-data data frame.}

\item{include_combos}{logical indicating whether or not to include -C tags in
the output - you will need to be careful with the drug_specifiers if T, e.g., 
\code{drug_specifiers = c("SCAN_ID", "lib_drug", "anchor")}}
}
\value{
data frame with columns \code{DRUG_ID, CELL_LINE_NAME, CL, maxc, x,
 y, drug, SCAN_ID, norm_neg_pos, drug_spec}
 \item{\code{CL}}{is the cell line identifier in the nlme model.}
 \item{\code{x}}{is the conversion of the micromolar screening concentration to a scale where the maximum dose = 9}
 \item{\code{y}}{is the normalized intensity from the experiment (sometimes called viability).}
 \item{\code{drug}}{is a concatenation of the drug_specifier columns (default DRUG_ID and maxc).}
 \item{\code{SCAN_ID}}{is the id of the plate scan used to provide the raw data.}
 \item{\code{norm_neg_pos}}{are the identifiers of the control tags used to normalise the data, separated by '+'.}
 \item{\code{drug_spec}}{is/are the column name(s) used to make the drug column separated by '+'. The original columns will be also included as separate columns.}
}
\description{
Converts GDSC normalized-data data frame to input format for nlme curve fit.
}
\examples{
data("gdsc_example")
gdsc_example <- removeFailedDrugs(gdsc_example)
gdsc_example <- removeMissingDrugs(gdsc_example)
gdsc_example <- normalizeData(gdsc_example)
gdsc_example <- setConcsForNlme(gdsc_example)
nlme_data <- prepNlmeData(gdsc_example, 
                          cl_id = "COSMIC_ID",
                           drug_specifiers = c("DRUGSET_ID", "lib_drug"))


}
\seealso{
\code{\link{removeFailedDrugs}},  \code{\link{removeMissingDrugs}},
  \code{\link{normalizeData}},  \code{\link{setConcsForNlme}}, 
  \code{\link{fitModelNlmeData}}
}
