% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit_prep.R
\name{removeMissingDrugs}
\alias{removeMissingDrugs}
\title{Removes library drugs listed with a drug id of NA from GDSC raw data.}
\usage{
removeMissingDrugs(myDat)
}
\arguments{
\item{myDat}{a GDSC raw data data frame.}
}
\description{
In contrast to drugs with 'FAIL' tags, the drugset contained NA for 
 the DRUG_ID, i.e., no drug id was assigned for that tag in the experimental
 design.
}
\details{
\code{removeMissingDrugs} removes rows from GDSC raw data where the
 \code{DRUG_ID} is NA.
}
\examples{
data("gdsc_example")
gdsc_example <- removeFailedDrugs(gdsc_example)
gdsc_example <- removeMissingDrugs(gdsc_example)
gdsc_example <- normalizeData(gdsc_example)
gdsc_example <- setConcsForNlme(gdsc_example)
nlme_data <- prepNlmeData(gdsc_example, "COSMIC_ID")

}
\seealso{
\code{\link{removeFailedDrugs}},  \code{\link{normalizeData}},
  \code{\link{setConcsForNlme}},  \code{\link{prepNlmeData}}
}
