% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit_prep.R
\name{set_package_used}
\alias{set_package_used}
\title{Get the package name and version information}
\usage{
set_package_used()
}
\value{
package name and version used to prepare the data as a string.
}
\description{
Get the package name and version information
}
