% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit_prep.R
\name{setConcsForNlme}
\alias{setConcsForNlme}
\title{Adds columns \code{maxc} and \code{x} to GDSC normalized-data data frame.}
\usage{
setConcsForNlme(
  normalized_data,
  group_conc_ranges = F,
  cell_line_spec = "MASTER_CELL_ID",
  conc_col = "CONC"
)
}
\arguments{
\item{normalized_data}{a GDSC normalized-data data frame.}

\item{group_conc_ranges}{logical. If TRUE then instances where the same drug
has been used in different concentration ranges will be grouped together
 with a single maxc. Default is FALSE.}

\item{cell_line_spec}{Column name to use for cell liene grouping. Default is MASTER_CELL_ID",}

\item{conc_col}{a string to identify the column used for the concentration 
values - useful for combination drug treatments.}
}
\description{
This is run prior to running prepNlmeData. The functionality is separate 
because the resulting `maxc` column might be used as a `drug_specifier`
 in prepNlmeData.
}
\details{
\code{maxc} is the maximum micromolar concentration of the treatment drug.
\code{x} is the conversion of the micromolar screening concentration to 
a range where the maximum dose = 9
}
\examples{
data("gdsc_example")
gdsc_example <- removeFailedDrugs(gdsc_example)
gdsc_example <- removeMissingDrugs(gdsc_example)
gdsc_example <- normalizeData(gdsc_example)
gdsc_example <- setConcsForNlme(gdsc_example)
nlme_data <- prepNlmeData(gdsc_example, "COSMIC_ID")

}
\seealso{
\code{\link{removeFailedDrugs}},  \code{\link{removeMissingDrugs}},
  \code{\link{normalizeData}},  \code{\link{prepNlmeData}}
}
