% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlme_fit_prep.R
\name{normalizeData}
\alias{normalizeData}
\title{Normalizes GDSC raw data intensities with respect to controls.}
\usage{
normalizeData(myDat, trim = T, neg_control = "NC-1",
  pos_control = "B")
}
\arguments{
\item{myDat}{a GDSC raw data data frame.}

\item{trim}{logical indicating whether to trim normalized values to the range
0 to 1. default \code{(trim = T)}}

\item{neg_control}{The tag used to recognise a negative control well - the 
upper end of the dynamic range.}

\item{pos_control}{The tag used to recognise a positive control well - the 
lower end of the dynamic range.}
}
\description{
\code{normalizeData} returns normalized intensities for the drug treated
wells - column \code{normalized_intensity}.
}
\details{
Replaces \code{TAG} column with columns: \code{lib_drug}, \code{dose} , and 
 \code{treatment} (single or combination - S or C respectively).
  
If anchor drug treatments (anchor + library combinations) exist in the data 
  then additional columns are added: \code{DRUG_ID_anch}, \code{CONC_anch}
  and \code{anchor}. 
  
Combination treatments (C) are treated as if the library alone is to be 
  fitted.
}
\examples{
data("gdsc_example")
gdsc_example <- removeFailedDrugs(gdsc_example)
gdsc_example <- removeMissingDrugs(gdsc_example)
gdsc_example <- normalizeData(gdsc_example)
gdsc_example <- setConcsForNlme(gdsc_example)
nlme_data <- prepNlmeData(gdsc_example, "COSMIC_ID")

}
\seealso{
\code{\link{removeFailedDrugs}},  \code{\link{removeMissingDrugs}},
  \code{\link{setConcsForNlme}},  \code{\link{prepNlmeData}}
}
